+++ test/ruby/test_regexp.rb
@@ -628,6 +628,12 @@ class TestRegexp < Test::Unit::TestCase
     assert_raise(RegexpError) { Regexp.new("((?<v>))\\g<0>") }
   end
 
+  def test_initialize_bool_warning
+    assert_warning(/expected true or false as ignorecase/) do
+      Regexp.new("foo", :i)
+    end
+  end
+
   def test_match_control_meta_escape
     assert_equal(0, /\c\xFF/ =~ "\c\xFF")
     assert_equal(0, /\c\M-\xFF/ =~ "\c\M-\xFF")